PROJ-44_signUp refactor: move sign up logic in custom hook
